CORDA-3989 Remove from `sessionsToBeClosed` when closing session
When closing a session manually, remove the session from both the open
sessions, as well as the sessions already marked as to be closed.

This prevents `findSessionsToBeTerminated` from finding a session that
has already been closed by `closeSessionTransition` that could have
executed before it.

Added logging of transition errors to `TopLevelTransition` to properly
log the fact that the error occurred while creating a transition as well
what event triggered the transition.