CORDA-3989 Terminate sessions instantly
Terminate sessions that need to be removed instantly in whatever
transition is currently executing, rather than scheduling another event
and doing so at a later time.

To do this, update the transition being created in `TopLevelTransition`
to remove the sessions and append the `RemoveSessionBindings` action to
it.

This achieves the same outcome as the original code but does so with 1
less transition. Doing this also removes the race condition that can
occur where another external event is added to the flow's event queue
before the terminate event could be added.