NOTICK Create a transition interceptor for testing
Define a corda service (annotated with `@CordaService`) which implements
`CustomTransitionInterceptor`.

Specify in the node configuration that this is the interceptor you want
to add.

Because of how it is implemented, you can effect the fiber itself but
you cannot change the outcome of the transition (other than throwing an
exception).