POC Await flow transitions from outside node
Add `await` + `release` functions to node to be able to wait for a flow
to enter a input status. Once it does, it decreases the permits in the
flow's semaphore, preventing the flow from continuing (blocks the flow
thread).

The permit is decreased inside of `LockingInterceptor`, which checks the
output state of a transition to determine whether to decrease the flow
locks permit count.

`release` calls `Semaphore.release` to put the lock back into a
uncontested state so that the flow can do the next transition.

`acquire` was not used, because it would block the calling thread. This
would decrease the usefulness in tests, as the test thread would be
blocked here.

A `Future` is returned by `await`, allowing the calling thread to do
other things, and then call `get` to block itself when it needs wait for
flow to reach the desired status.

Currently this is hard coded into the node, but it could be changed to
only be enabled in dev mode or something like that. It is accessed via
the `MultiRPCClient`.

I'll tidy it up if we are happy with the idea. A.k.a, remove the
reflection and not shove everything into
`SingleThreadedStateMachineManager`.