ENT-5750 Keep killed flows that were started with client ids
If a flow is started with a client id, do not delete it from the
database. Instead, set the status to `KILLED` and store a
`KilledFlowException` in the database.

Keeping it around allows it to be reattached to using the same
mechanisms as active, completed or failed flows. Furthermore, without
this change it could be possible for a flow to be triggered again if
killed while the reconnecting rpc client is being used.

If there is no client id related to the flow, then kill flow keeps its
original behaviour of deleting all traces of the flow.