ENT-5750 Cannot kill finished flows
Flows cannot be killed if they are `COMPLETED`, `FAILED` or `KILLED`
already.

Logs have been added if requests to kill flows in these statuses are
made.

Do not update the status + persist the exception if the flow was killed
during flow initialisation (before persisting its first checkpoint).

Remove the client id mapping if the flow was killed and did not persist
its original checkpoint.