ENT-6376 Cancellation of `ReceiveFinalityFlow`
Add the ability for a peer to cancel any peers that are waiting to
receive a `SignedTransaction` within `ReceiveFinalityFlow` and
consequently within `ReceiveTransactionFlow`.

This is to allow flows to catch errors, such as `NotaryException` and
handle what happens next in a graceful manner. For example, catching the
`NotaryException` and gracefully terminate the peers because the CorDapp
expected this to happen.

Cancellation of peers is achieved by calling
`FinalityFlow.cancelPeerReceiveTransactionFlows` on the peer that called
`FinalityFlow`.

This sends an `InvalidCoreTransaction` contained in a
`SignedTransaction` to peers. The sending of a `SignedTransaction` is
used to wake up the peers in `ReceiveTransactionFlow` as they're waiting
for a transaction to be sent. The flow then checks whether
`SignedTransaction.coreTransaction` is an `InvalidCoreTransaction`. If
so, it throws a `CancelPeerReceiveTransactionFlowException` that the
hospital allows to propagate, so that it can be caught by the calling
user coded flow.

When an `InvalidCoreTransaction` is sent to a node running an older
version of Corda without this change, it will experience a serialization
error and leads to the flow being hospitalized. This is the same
behaviour before the change was made, however the error is more
confusing.

This method was used, because we cannot send an exception to peers
without also terminating the sending flow.