ENT-6286: More fixes to dynamic user population logic
If the user is removed mid-session, given that Artemis cache is now disabled, it will be Artemis (and not Shiro) rejecting the user.

To receive response from Artemis we need to switch to blocking non-durable send mode on the RPC Client side. Furthermore, once Artemis rejected the user - the producer associated with the session becomes closed and can no longer be used for anything, e.g. closing the `Observables` so this has to be done in the lenient fashion.