ENT-9569: Apply the 60s SSL handshake timeout to the embedded Artemis server
This ensures that any CRL connection timeouts due to unreachable dist points will still be captured by the soft fail flag if it’s enabled. Otherwise, the default 10s SSL handshake timeout is too quick against the default 15s CRL timeout in the JDK. This was achieved by using a custom `AcceptorFactory` (`NodeNettyAcceptorFactory`) which sets the `handshakeTimeoutMillis` property on the Netty `SslHandler`.

A series of other change were made to facilate this change:
* `sslHandshakeTimeout` is now a `Duration` rather than long millis, which is helpful since the JDK CRL timeout is in seconds
* Any exceptions ignored due to soft fail are logged as debug
* The new acceptor factory also adds the Netty logging handler if `trace` is true
* The transport methods in `ArtemisTcpTransport` have been cleaned up and duplication removed. It’s now easier to understand which transports need to the SSL handshake timeout parameter
* `CertificationRevocationListNodeTests` has been cleaned up and duplication removed. It’s now easier to understand the various test cases. SSL handshake timeout tests were added for both the Artemis and the AMQP server classes
* lib/quasar.jar, which is meant to be git ignored, has been removed from git tracking