ENT-13924 - Artemis upgrade, preserve Caffeine (#7974)
Upgraded Artemis to get past numerous security issues.
A consequence of this is that Artemis upgrades the dependency on
Caffeine from 3.1.8 to 3.2.1, which causes a plethora of compiler
warnings.
The changes presented here forcibly keep Caffeine at 3.1.8, except for
the legacy sub-projects which require a JDK8-compatible version of
Caffeine.