ENT-13929: Notary instruction component group (#7976)
Introducing a new transaction component group, notary instructions.

These instructions are represented by the marker interface
`NotaryInstruction` and represent arbitrary instructions that notaries
can perform alongside the notarisation of the transaction. This is not a
plugin mechanism, and only R3 will provide implementations either in
Core or via extension SDKs. For now the notary implementations do not
support instructions, and in fact expect no instructions to be present.
This later point is to prevent unknown instructions being used.

Notary instructions are added via
`TransactionBuilder.addNotaryInstruction` and can be queried in
`LedgerTransaction`s. This can be used in contract verification to
mandate certain instructions are present alongside the Corda ledger
state transition. Filtered transactions also have this component group,
and are included when sent to non-validating notaries.

There's also some cleanup, including back-porting changes made in
`UniquenessProviderTests` in ENT.