ENT-14451: Ensure that a CorDapp is present in cordapps directory when choosing an attachment for a missing class dependency (#8014)
Two separate problems were flagged by @Tagada14 :
1. `When building a transaction, the TransactionBuilder attempts to
verify a transaction, and if it encounters a missing class dependency,
it looks for it in the attachments and adds the attachment with the
missing class to the transaction. However, it considers all of the
attachments from the Attachment Storage rather than the CorDapps present
in the node. This allows it to add an arbitrary trusted attachment that
contains the missing class, even a legacy one!`
**Fix:**
The attachment lookup logic has been refactored to prioritise installed
CorDapps before falling back to the database.
**- Installed-first logic:**
getTrustedClassAttachments now searches the local CorDapps directories
first. If no suitable attachment is found locally, it may optionally
fall back to the database.
**- Legacy handling:**
Legacy contracts only search in legacy-contracts/ and never fall back to
the database.
**- JDK 17 filtering:**
When searching in the DB, only attachments compatible with JDK 17 are
considered.
**Two new system properties:**
- `net.corda.node.transactionbuilder.missingClassDbSearchDisabled`
(default false) - If `true`, disables database fallback. Only installed
CorDapps will be searched.
- `net.corda.node.transactionbuilder.installedFirstSearchDisabled`
(default false) - If `true`, disables the new “installed-first” lookup
logic and reverts to the old database-only behaviour.

2. `Moreover, it doesn't check if multiple attachments contain said
class; it simply chooses the first one. Because of this, the attachment
it chooses is runtime-dependent and non-deterministic. With the same set
of attachments in the storage, one node may attach CorDapp A with the
missing class and the other CorDapp B.`
This has been fixed by improving the sorting mechanism when returning
attachments - now attachments are additionally (apart from primary
sorting by version descending) sorted by their IDs, which ensures
deterministic order of returned attachments. For example, if we have two
attachments with the same version, they'll be ordered alphabetically,
not returned in random order as before.

**Credits to @Tagada14 for finding the above bug and suggesting a good
solution in https://github.com/corda/corda/pull/7994.**

# PR Checklist:

- [ ] Have you run the unit, integration and smoke tests as described
[here](https://docs.r3.com/testing.html)?
- [ ] If you added public APIs, did you write the JavaDocs/kdocs?
- [ ] If the changes are of interest to application developers, have you
added them to the changelog, and potentially the [release
notes](https://docs.r3.com/release-notes.html)
(`https://docs.r3.com/release-notes.html`)?
- [ ] If you are contributing for the first time, please read the
[contributor agreement](https://docs.r3.com/contributing.html) now and
add a comment to this pull request stating that your PR is in accordance
with the [Developer's Certificate of
Origin](https://docs.r3.com/contributing.html).

Thanks for your code, it's appreciated! :)

---------

Co-authored-by: Adel El-Beik <adel.el-beik@r3.com>