ENT-14093: Backporting changes to `UniquenessProviderTests` (#8061)
https://github.com/corda/enterprise/pull/5579 improved the assertions in
`UniquenessProviderTests`, and also had to slightly change some of the
tests which assumed the `UniquenessProvider` will always process
notarisation requests in the same order they are submitted. This
assumption only works if the implementation uses a singleton executor,
which most of them do, but not the Solana one which uses a thread pool.